It checks if the user user_0d_1 already exists in the MySQL server by querying the mysql.user table. If the user exists, it sets the variable @user_exists to 1; otherwise, it sets it to 0.
If the user does not exist (@user_exists is 0), it creates the user user_0d_1 with the specified password user_0d_1_pwd.

It grants all privileges on all databases and tables (*.*) to user_0d_1 and includes the WITH GRANT OPTION to allow the user to grant privileges to other users. This effectively gives the user full control over the MySQL server.

After creating the user and granting privileges, it flushes privileges to make the changes take effect.

It provides a message indicating whether the user was created or if the user already existed.